feat: move rf log from debug page to syslog page and save log level in config

+++ software/NRG_itho_wifi/main/generic_functions.h
@@ -42,7 +42,7 @@ void logLastCommand(const char *command, const char *source);
 void getLastCMDinfoJSON(JsonObject root);
 void updateItho();
 void add2queue();
-void setRFdebugLevel(uint16_t level);
+void setRFdebugLevel(uint8_t level);
 double round(float value, int precision);
 char toHex(uint8_t c);
 std::vector<int> parseHexString(const std::string &input);
